+++ src/app/item-feed/ItemFeed.tsx
@@ -114,7 +114,7 @@ export default function ItemFeed({
             type="button"
             className={clsx('dim-button', styles.clearButton)}
             onClick={() => {
-              setItemFeedWatermark(undefined);
+              setItemFeedWatermark('0');
               // Don't spawn all the items at the same time again
               resetItemCount();
             }}