+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * If you add hashtags to your loadouts' names or notes, DIM will show buttons for quickly filtering down to loadouts that include that hashtag.
+* Fixed a bug where the "Show Older Items" button in the Item Feed would not permanently show all old items.
 
 ## 7.50.3 <span class="changelog-date">(2023-01-04)</span>